# app/swagger.yml
swagger: "2.0"
info:
  description: This is the swagger file that goes with our server code
  version: "1.0.0"
  title: Swagger REST Article
consumes:
  - application/json
produces:
  - application/json
basePath: /sample_spa/v1
...
paths:
...
  /student/{usn}:
    get:
      operationId: app.student.read_one
      tags:
        - student
      summary: Read one student from the student list
      description: Read one student from the student list
      parameters:
        - name: usn
          in: path
          description: Last name of the student to get from the list
          type: string
          required: True
      responses:
        200:
          description: Successfully read student from student list operation
          schema:
            properties:
              fname:
                type: string
              lname:
                type: string

    put:
      operationId: app.student.update
      tags:
        - student
      summary: Update a student in the student list
      description: Update a student in the student list
      parameters:
        - name: usn
          in: path
          description: Last name of the student to update in the list
          type: integer
          required: True
        - name: student
          in: body
          schema:
            type: object
            properties:
              fname:
                type: string
              lname:
                type: string
      responses:
        200:
          description: Successfully updated student in student list

    delete:
      operationId: app.student.delete
      tags:
        - student
      summary: Delete a student from the student list
      description: Delete a student
      parameters:
        - name: usn
          in: path
          type: integer
          required: True
      responses:
        200:
          description: Successfully deleted a student from student list